% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ctpp.R
\name{download_ctpp}
\alias{download_ctpp}
\title{Download CTPP Data}
\usage{
download_ctpp(id, geography = "03", state = "", dataset = "",
  output = "Name", debug_request = F)
}
\arguments{
\item{id}{The identifier of the CTPP table you want to download (e.g. "A101100"). Type \link[CTPPr]{ctpp_tables}() for a list of tables and their ids.}

\item{geography}{The geography level of the requested table. Type ctpp_geography_list to see available geographies, or see Details below.}

\item{state}{The state your request should subset to. Takes name ("Alabama"), abbreviation ("AL") or FIPS ("01"). Type ctpp_state_list to see available states.}

\item{dataset}{"2010" or "2016" to specify which 5-year dataset to query.}

\item{output}{How you want your records labelled. Default is "Name" unless geography is "Tract", "TAD", or "TAZ", in which case "FIPS Code" is used to reduce size of request. See Details below for example outputs.}
}
\value{
returns a data.table based on a default option CSV download request (tables available at the \href{http://data5.ctpp.transportation.org/ctpp/Browse/BrowseTables.aspx}{online CTPP browser})
}
\description{
This function allows you to download CTPP tables into R
}
\section{Details}{


This function makes a bulk download request for the provided table id. Use this function with caution and subset to a specific state for smaller geography requests. See the example section for some recommendations on downloading and storing your results.

\subsection{geography}{
{Bold words below are accepted by geography=}
{For Residence and Workplace Tables:}
  \itemize{
    \item \strong{County} - State-County
    \item \strong{MCD} - State-County-MCD (for 12 strong MCD states)
    \item \strong{Place} - State-Place
    \item \strong{MSA} - Metropolitan Statistical Area
    \item \strong{City} - Metropolitan Statistical Area - EACH Principal City
    \item \strong{PUMA} - State-PUMA5
    \item \strong{UA} - Urbanized Area (UA)
    \item \strong{OUSA} - Worked Outside United States (WORKPLACE tables only)
    \item \strong{Tract} - State-County-Tract
    \item \strong{TAD} - TAD
    \item \strong{TAZ} - TAZ
  }
{For Flow (Residence->Workplace) Tables:}
  \itemize{
    \item \strong{State->State}
    \item \strong{County->County}
    \item \strong{MCD->MCD}
    \item \strong{Place->Place}
    \item \strong{City->City}
    \item \strong{PUMA->PUMA}
    \item \strong{County->Place}
    \item \strong{MCD->Place}
    \item \strong{PUMA->Place}
    \item \strong{Tract->Tract}
    \item \strong{TAD->TAD}
    \item \strong{TAD->TAZ}
    \item \strong{TAZ->TAZ}
    \item \strong{TAZ->TAD}
    \item \strong{Place->TAZ}
    \item \strong{TAZ->Place}
  }
}
\subsection{state}{
{Use state= in combination with geography=. If you don't, the requested table will be downloaded for the entire country and may not finish.}
}
\subsection{output}{
{\strong{Option} - Example}
  \itemize{
    \item \strong{Name} - "New York"
    \item \strong{Key} - "C0300US13001"
    \item \strong{FIPS Code} - "36"
    \item \strong{FIPS and Name} - "36 | New York"
    \item \strong{Split FIPS Code} - "36 | 001"
  }
}
}

\examples{
\donttest{
A102102 <- download_ctpp(
  id = "A102102",
  geography = "County",
  state = "GA"
)
A202100 <- download_ctpp(
  id = "A202100",
  geography = "MSA",
  state = "FL"
)
# save your tables for working with later
save(A102102, A202100, "CTPP-tables.RData")
# load tables back
load("CTPP-tables.RData")
}
}
